+++ pilco/__init__.py
@@ -1,5 +1,5 @@
 from . import models
-# from . import controllers
-# from . import rewards
-# from . import utils
-# from . import safe_pilco
+from . import controllers
+from . import rewards
+from . import utils
+from . import safe_pilco
